Move common rules functions to rules_common
